requirements.txt
colorama==0.4.3
Pillow==7.0.0
pygooglechart==0.4.0
python-barcode==0.11.0
qrcode==6.1
six==1.14.0
